avec plots

+++ src/treatment_pipeline.py
@@ -31,8 +31,7 @@ class Treatment():
     def __init__(self, args):
         # TODO replace subjects from 1 to 109 and runs to 0, 4:
         self.subjects = [args.subject] if args.train or args.predict else range(1, 109)
-        # self.run_idxs = [args.run_idx] if args.train or args.predict else range(0, 4)
-        self.run_idxs = [args.run_idx] if args.train or args.predict else [1, 3]
+        self.run_idxs = [args.run_idx] if args.train or args.predict else range(0, 4)
         self.runs = [x['runs'] for i, x in enumerate(tasks) if i in self.run_idxs]
         self.my_own_csp = args.my_csp
 
@@ -41,7 +40,6 @@ class Treatment():
             lda = LDA(solver='lsqr', shrinkage='auto')
         elif lda_name == 'classic':
             lda = LDA()
-        # TODO: Make sure it works:
         if self.my_own_csp:
             print("Using my own CSP")
             csp = My_CSP(n_components=csp_components)
@@ -71,13 +69,12 @@ class Train(Treatment):
         super().__init__(args)
         self.training_data = None
         self.pipe = None
-        self.preload_models = False
+        self.preload_models = True
         self.train()
 
     def cross_val_pipeline(self, pipeline, epochs) -> float:
         targets = epochs.events[:, -1]
         epochs_data = epochs.copy().crop(tmin=1.0, tmax=4.0).get_data() # They do this cropping on MNE website. Why ?
-        # TODO: keep or remove random state?
         scores = cross_val_score(pipeline, epochs_data, targets, cv=ShuffleSplit(10, test_size=0.2, random_state=42), n_jobs=None)
         return scores.mean()
 
@@ -90,19 +87,18 @@ class Train(Treatment):
 
         x_train, y_train = epochs_data, targets
         pipeline.fit(X=x_train, y=y_train)
-        # TODO: put it back to test!!
-        # score = pipeline.score(X=x_test, y=y_test)
-        score = pipeline.score(X=x_train, y=y_train)
+        score = pipeline.score(X=x_test, y=y_test)
         logger.info(f"Training score: {self.color_good(score, (score >= 0.75))}")
         return score
 
-    def save_pipeline_to_disk(self, pipe: Pipeline, dir: str, run_id: int, subject: int, parser):
+    def save_pipeline_to_disk(self, pipe: Pipeline, dir: str, run_id: int, subject: int):
         sub_dir = f"run_{run_id}"
-        parser.create_dir_if_not_exists(dir, sub_dir)
+        Parser.create_dir_if_not_exists(dir, sub_dir)
         filename = f"S{subject:03d}.save"
         complete_path = f'{dir}/{sub_dir}/{filename}'
-        joblib.dump(pipe, complete_path)
-        logger.info(f"Saved pipeline to: {complete_path}")
+        if not os.path.exists(complete_path):
+            joblib.dump(pipe, complete_path)
+            logger.info(f"Saved pipeline to: {complete_path}")
 
     def train(self):
         training_data = pd.DataFrame(columns=['subject', 'task_number', 'cross_val_score', 'training_score'])
@@ -113,20 +109,19 @@ class Train(Treatment):
             for subject in self.subjects:
                 parser = Parser(subject=subject, run=run, mne_path="./mne_data", run_id=run_id)
                 parser.motion_preprocessing(labels)
-                # TODO make it a class method:
                 parser.create_dir_if_not_exists("./results")
                 epochs = parser.get_epochs(epochs_dir="./epochs", save=True, preload=True)
-                cross_score = None
                 if self.preload_models:
                     self.pipe = self.get_pretrained_model("./models", run_id, subject)
                 if self.pipe is None or self.preload_models is False:
-                    self.pipe = self.build_pipe(lda_name='shrimp', csp_components=6)
-                    cross_score = self.cross_val_pipeline(self.pipe, epochs)
+                    self.pipe = self.build_pipe(lda_name='shrimp', csp_components=10)
+                cross_score = self.cross_val_pipeline(self.pipe, epochs)
+                if cross_score:
+                    logger.info(f"Run:{run},\tsubject:{subject},\tCross validation average:{cross_score:.2f}")
+                else:
+                    logger.info(f"Run:{run},\tsubject:{subject}")
                 trained_pipeline_score = self.train_pipeline(self.pipe, epochs, test_size=0.1)
-                # self.save_pipeline_to_disk(self.pipe, "./models", run_id, subject, parser)
-                # TODO: put cross val before train:
-                cross_score = round(cross_score, 2) if type(cross_score) == float else None
-                logger.info(f"Run:{run},\t\tsubject:{subject},\t\tCross validation average:{cross_score}")
+                self.save_pipeline_to_disk(self.pipe, "./models", run_id, subject)
                 model_metrics = {
                     'subject': subject,
                     'task_number': run_id,
